Moving files, updating namespaces, removing unused usings
