+++ eShopModernizedMVCSolution/src/eShopModernizedMVC/App_Start/BundleConfig.cs
@@ -1,5 +1,4 @@
-﻿using System.Web;
-using System.Web.Optimization;
+﻿using System.Web.Optimization;
 
 namespace eShopModernizedMVC
 {
