+++ eShopModernizedMVCSolution/src/eShopModernizedMVC/App_Start/CatalogConfiguration.cs
