+++ eShopModernizedMVCSolution/src/eShopModernizedMVC/App_Start/FilterConfig.cs
@@ -1,5 +1,4 @@
 ﻿using eShopModernizedMVC.Filters;
-using System.Web;
 using System.Web.Mvc;
 
 namespace eShopModernizedMVC
