+++ eShopModernizedMVCSolution/src/eShopModernizedMVC/App_Start/MyTelemetryInitializer.cs
