+++ eShopModernizedMVCSolution/src/eShopModernizedMVC/Startup.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
+﻿using eShopModernizedMVC.Middleware;
 using Microsoft.Owin;
 using Owin;
